LayoutTests/svg/custom/svg-parse-overflow-5-expected.txt

CONSOLE MESSAGE: line 8: Error: Invalid value for <foreignObject> attribute y="105.0e5000000000000000000000000000000000000000000000000000"
